+++ linen-modern/src/main/scala/x7c1/linen/modern/accessor/SourceStateAccessor.scala
@@ -1,4 +1,4 @@
-package x7c1.linen.modern
+package x7c1.linen.modern.accessor
 
 trait SourceStateAccessor {
   def findState(sourceId: Long): Option[SourceState]
