+++ linen-modern/src/main/scala/x7c1/linen/modern/action/observer/EntryFocusedObserver.scala
@@ -1,6 +1,7 @@
-package x7c1.linen.modern
+package x7c1.linen.modern.action.observer
 
-import x7c1.linen.modern.CallbackTaskRunner.runAsync
+import x7c1.linen.modern.action.observer.CallbackTaskRunner.runAsync
+import x7c1.linen.modern.action.{Actions, EntryFocusedEvent}
 import x7c1.wheat.macros.logger.Log
 import x7c1.wheat.modern.observer.OnItemFocusedListener
