+++ linen-modern/src/main/scala/x7c1/linen/modern/display/EntryDetailRowAdapter.scala
@@ -1,10 +1,11 @@
-package x7c1.linen.modern
+package x7c1.linen.modern.display
 
 import android.support.v7.widget.RecyclerView.Adapter
 import android.view.ViewGroup
 import x7c1.linen.glue.res.layout.EntryDetailRow
+import x7c1.linen.modern.accessor.EntryAccessor
+import x7c1.linen.modern.struct.Entry
 import x7c1.wheat.ancient.resource.ViewHolderProvider
-import x7c1.wheat.macros.logger.Log
 import x7c1.wheat.modern.decorator.Imports._
 
 
@@ -39,18 +40,4 @@ case class EntryDetailSelectedEvent(position: Int, entry: Entry){
   def dump: String = s"position:$position, entry:$entry"
 }
 
-class EntryDetailSelectedObserver (actions: Actions)
-  extends OnEntryDetailSelectedListener {
 
-  import x7c1.linen.modern.CallbackTaskRunner.runAsync
-
-  override def onEntryDetailSelected(event: EntryDetailSelectedEvent): Unit = {
-    Log info s"[init] ${event.dump}"
-
-    val sync = for {
-      _ <- actions.detailArea.onEntryDetailSelected(event)
-    } yield ()
-
-    Seq(sync) foreach runAsync { Log error _.toString }
-  }
-}
