+++ linen-modern/src/main/scala/x7c1/linen/modern/display/EntryRowAdapter.scala
@@ -1,8 +1,10 @@
-package x7c1.linen.modern
+package x7c1.linen.modern.display
 
 import android.support.v7.widget.RecyclerView.Adapter
 import android.view.ViewGroup
 import x7c1.linen.glue.res.layout.EntryRow
+import x7c1.linen.modern.accessor.EntryAccessor
+import x7c1.linen.modern.struct.Entry
 import x7c1.wheat.ancient.resource.ViewHolderProvider
 import x7c1.wheat.modern.decorator.Imports._
 
@@ -35,9 +37,6 @@ trait OnEntrySelectedListener {
   def onEntrySelected(event: EntrySelectedEvent): Unit
 }
 
-case class EntrySelectedEvent(
-  position: Int,
-  entry: Entry ){
-
+case class EntrySelectedEvent(position: Int, entry: Entry){
   def dump: String = s"entryId:${entry.entryId}, position:$position"
 }
