+++ linen-modern/src/main/scala/x7c1/linen/modern/display/SourceRowAdapter.scala
@@ -1,8 +1,10 @@
-package x7c1.linen.modern
+package x7c1.linen.modern.display
 
 import android.support.v7.widget.RecyclerView.Adapter
 import android.view.{View, ViewGroup}
 import x7c1.linen.glue.res.layout.SourceRow
+import x7c1.linen.modern.accessor.{SourceAccessor, SourcePrefetched, SourceStateAccessor}
+import x7c1.linen.modern.struct.Source
 import x7c1.wheat.ancient.resource.ViewHolderProvider
 import x7c1.wheat.modern.decorator.Imports._
 
@@ -22,7 +24,7 @@ class SourceRowAdapter(
     holder.title.text = source.title
     holder.description.text = source.description
     holder.itemView onClick { view =>
-      val event = SourceSelectedEvent(source, position)
+      val event = SourceSelectedEvent(position, source)
       sourceSelectedListener onSourceSelected event
     }
     holder.statePrefetched.setVisibility(View.GONE)
@@ -41,4 +43,8 @@ class SourceRowAdapter(
 
 trait OnSourceSelectedListener {
   def onSourceSelected(event: SourceSelectedEvent): Unit
+}
+
+case class SourceSelectedEvent (position: Int, source: Source){
+  def dump: String = s"sourceId:${source.id}, position:$position"
 }
