+++ linen-modern/src/main/scala/x7c1/linen/modern/struct/Entry.scala
@@ -1,0 +1,10 @@
+package x7c1.linen.modern.struct
+
+case class Entry(
+  sourceId: Long,
+  entryId: Long,
+  url: String,
+  title: String,
+  content: String,
+  createdAt: Date
+)
