+++ linen-modern/src/main/scala/x7c1/linen/modern/struct/Source.scala
@@ -1,0 +1,8 @@
+package x7c1.linen.modern.struct
+
+case class Source(
+  id: Long,
+  url: String,
+  title: String,
+  description: String
+)
